+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - DatePicker: Added tests for max, min attributes and onEnterPress event
 - Checkbox: Added tests checking that className attribute is passed to Wrapper, not LabelElement
 - DatePicker: Added possibility to pass custom month and week day name through props. New props: monthNames, shortMonthNames, weekDayNames, shortWeekDayNames.
+- DatePicker: Added outputFormat prop. Pass output date string format to get formatted value. Example: 'dd.MM.yyyy'. Formatted value can be obtained by handlers: onBlur, onChange, onEnterPress, onFocus.
 
 ### Fixed
 - Input: Focus input on clear button click.